scraper/requirements.txt
attrs==19.3.0
Automat==0.8.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
dateutils==0.6.7
feedparser==5.2.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==17.5.0
lxml==4.5.0
more-itertools==8.2.0
packaging==20.1
parsel==1.5.2
pluggy==0.13.1
Protego==0.1.16
psycopg2-binary==2.8.4
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==2.0.0
pymongo==3.10.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
pytz==2019.3
queuelib==1.5.0
requests==2.22.0
Scrapy==1.8.0
service-identity==18.1.0
six==1.14.0
soupsieve==1.9.5
Twisted==19.10.0
urllib3==1.25.8
w3lib==1.21.0
wcwidth==0.1.8
zipp==2.2.0
zope.interface==4.7.1
